+++ crates/lune-std-net/src/shared/mod.rs
@@ -1,4 +1,3 @@
-pub mod body;
 pub mod futures;
 pub mod headers;
 pub mod hyper;